refactor(db): Move db to src/db from src/lib/db
